feat(email): add PDF attachment
