+++ frontend/src/components/FicheqsForm/FicheqsForm.tsx
@@ -3,6 +3,7 @@ import { useNavigate } from "react-router-dom";
 import { toInputDateValue } from "../../utils/date";
 import { getCookie } from "../../utils/cookie";
 import { formatStatusTag } from "../../utils/status";
+import { blobToBase64 } from "../../utils/blobToBase64";
 import { createFicheqs } from "../../services/api";
 import { sendPDF } from "../../services/api";
 import { generatePDF } from "../FicheqsPDF/FicheqsPDF";
@@ -915,7 +916,7 @@ const FicheqsForm: React.FC<FicheqsFormProps> = ({
         event.preventDefault();
 
         try {
-            // Rassemblez toutes les données des champs
+            // Tous les champs
             const allFields = [
                 ...fieldsDaaf,
                 ...fieldsGaz,
@@ -946,26 +947,22 @@ const FicheqsForm: React.FC<FicheqsFormProps> = ({
             // Génération du PDF
             const pdfBlob = generatePDF(pdfData);
 
-            // Demande de l'email du destinataire
-            const recipientEmail = prompt("Adresse email du destinataire :");
-            if (!recipientEmail) return;
+            // Conversion du blob PDF en base64
+            const attachmentBase64 = await blobToBase64(pdfBlob);
 
-            // Conversion du blob PDF en base64 pour l'envoi JSON
-            const base64PDF = await blobToBase64(pdfBlob);
-
-            // Création de l'objet à envoyer via l'API
+            // Création de l'objet pour l'API selon le format attendu dans email.routes.ts
             const emailData = {
-                to: recipientEmail,
+                to: email, // Utiliser l'email du cookie
                 subject: `FicheQS - ${logement}`,
                 body: `Veuillez trouver ci-joint la fiche qualité sécurité pour le logement ${logement}.`,
-                attachmentBase64: base64PDF,
+                attachmentBase64: attachmentBase64, // Envoi direct du base64, pas d'un tableau
             };
 
-            // Utilisation de votre service API
+            // Appel de la fonction de l'API qui gère l'envoi
             const result = await sendPDF(emailData);
 
             if (result.success) {
-                alert("Email envoyé avec succès !");
+                console.log("Email envoyé avec succès !");
             } else {
                 throw new Error(result.message || "Erreur lors de l'envoi de l'email");
             }
@@ -973,21 +970,6 @@ const FicheqsForm: React.FC<FicheqsFormProps> = ({
             console.error("Erreur lors de l'envoi de l'email:", error);
             alert("Une erreur est survenue lors de l'envoi de l'email");
         }
-    };
-
-    // Fonction utilitaire pour convertir un Blob en base64
-    const blobToBase64 = (blob: Blob): Promise<string> => {
-        return new Promise((resolve, reject) => {
-            const reader = new FileReader();
-            reader.onloadend = () => {
-                const base64String = reader.result as string;
-                // Extraire uniquement la partie base64 sans le préfixe "data:application/pdf;base64,"
-                const base64 = base64String.split(",")[1];
-                resolve(base64);
-            };
-            reader.onerror = reject;
-            reader.readAsDataURL(blob);
-        });
     };
 
     return (
